refactor: split chart and demo modules
