+++ examples/demo.rs
@@ -31,6 +31,10 @@ use tuika::framebuffer::{FrameBuffer, FrameBufferView};
 use tuika::prelude::*;
 use tuika::view::DrawView;
 
+#[path = "demo_scenes/mod.rs"]
+mod demo_scenes;
+use demo_scenes::*;
+
 /// A tiny self-contained Rust highlighter for the `markdown`/`code_block` scenes.
 ///
 /// The examples deliberately avoid a grammar dependency (that would drag
@@ -949,1216 +953,4 @@ fn run(d: &Demo) -> io::Result<()> {
     let _ = terminal.clear();
     drop(terminal);
     Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Scenes. Each is a pure function of the frame counter and theme.
-// ---------------------------------------------------------------------------
-
-fn labeled_row(glyph: Element, label: &str, theme: &Theme) -> Element {
-    let text = Text::new(vec![Line::from(Span::styled(
-        label.to_string(),
-        theme.text_style(),
-    ))]);
-    view! {
-        row(gap = 1) {
-            fixed(2) { node(glyph) }
-            grow(1) { node(text) }
-        }
-    }
-}
-
-fn scene_spinner(frame: u64, theme: &Theme) -> Element {
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(labeled_row(element(Spinner::new(frame).style(SpinnerStyle::Braille)), "Braille — the smooth default", theme)) }
-            fixed(1) { node(labeled_row(element(Spinner::new(frame).style(SpinnerStyle::Line)), "Line — ASCII fallback", theme)) }
-            fixed(1) { node(labeled_row(element(Spinner::new(frame).style(SpinnerStyle::Dots)), "Dots — bouncing", theme)) }
-        }
-    }
-}
-
-fn scene_progress(frame: u64, theme: &Theme) -> Element {
-    let animated = tuika::anim::ping_pong(frame, 140);
-    let _ = theme;
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(ProgressBar::determinate(0.25).percent(true)) }
-            fixed(1) { node(ProgressBar::determinate(0.60).label("0:42 / 3:07").percent(true)) }
-            fixed(1) { node(ProgressBar::determinate(animated).percent(true)) }
-            fixed(1) { node(ProgressBar::indeterminate(frame)) }
-        }
-    }
-}
-
-fn scene_activity_list(frame: u64, theme: &Theme) -> Element {
-    let _ = theme;
-    let progress = tuika::anim::ping_pong(frame, 180);
-    element(
-        ActivityList::new(vec![
-            ActivityItem::new("Resolve dependencies", ActivityStatus::Succeeded)
-                .detail("42 crates"),
-            ActivityItem::new("Compile workspace", ActivityStatus::Running)
-                .detail("tuika")
-                .progress(progress),
-            ActivityItem::new("Run tests", ActivityStatus::Queued),
-            ActivityItem::new("Publish artifacts", ActivityStatus::Skipped),
-        ])
-        .frame(frame)
-        .gap(1),
-    )
-}
-
-fn scene_loader(frame: u64, theme: &Theme) -> Element {
-    let _ = theme;
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(Loader::new(frame, "compiling crate…").hint("esc to cancel")) }
-            fixed(1) { node(Loader::new(frame, "fetching dependencies…").spinner_style(SpinnerStyle::Line)) }
-        }
-    }
-}
-
-fn scene_text(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let styled = Text::new(vec![
-        Line::from(vec![
-            Span::styled(
-                "error",
-                Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-            ),
-            Span::styled(": unresolved import ", theme.text_style()),
-            Span::styled("`tuika::Widget`", theme.accent_style()),
-        ]),
-        Line::from(Span::styled(
-            "  perhaps you meant `tuika::View`?",
-            theme.muted_style(),
-        )),
-    ]);
-    let prose = tuika::components::Paragraph::new(
-        "Paragraph word-wraps plain text to the render width in a single style, \
-         re-flowing every frame so a resize just re-wraps.",
-        theme.text_style(),
-    );
-    view! {
-        col(gap = 1) {
-            fixed(2) { node(styled) }
-            grow(1) { node(prose) }
-        }
-    }
-}
-
-/// The document the `markdown` scene streams in, one glyph at a time.
-const MARKDOWN_DOC: &str = "\
-# Streaming Markdown
-
-Renders **CommonMark** as it *streams* — only the in-flight tail
-re-parses, so settled blocks and `code` never re-tokenize.
-
-- headings, **bold**, *italic*, `inline code`
-- nested lists and tables
-
-```rust
-fn greet(name: &str) {
-    println!(\"hello, {name}!\");
-}
-```
-";
-
-/// Animated: reveal `MARKDOWN_DOC` progressively through a `MarkdownState`, the
-/// same way a host feeds an assistant message as it streams. Holds on the full
-/// document, then restarts.
-fn scene_markdown(frame: u64, theme: &Theme) -> Element {
-    let _ = theme;
-    let total = MARKDOWN_DOC.chars().count();
-    // Reveal briskly so the whole document — including the highlighted code
-    // block — lands within a short recording, then hold before the cycle repeats.
-    let pos = (frame as usize * 6) % (total + 120);
-    let revealed: String = MARKDOWN_DOC.chars().take(pos.min(total)).collect();
-
-    let mut state = MarkdownState::new();
-    state.set(revealed);
-    // Rendered at a fixed width for the demo frame; a real host passes the live
-    // viewport width so prose re-wraps on resize.
-    let sheet = tuika::StyleSheet::from_theme(theme);
-    let lines = state
-        .lines(
-            64,
-            theme,
-            &sheet,
-            tuika::highlight::CodeHighlighter::With(&HL),
-        )
-        .to_vec();
-    element(Text::new(lines))
-}
-
-/// The GFM table the `markdown_table` scene renders.
-///
-/// Deliberately exercises everything a table cell can carry: per-column
-/// alignment from the `:---:` markers, inline code and bold, wide emoji (which
-/// must be measured grapheme-aware or the borders drift), and links whose label
-/// is what gets painted while the destination rides along as OSC 8.
-const MARKDOWN_TABLE_DOC: &str = "\
-| Component   |  Status   |                          Docs |
-| :---------- | :-------: | ----------------------------: |
-| `Markdown`  | ✅ stable | [docs.rs](https://docs.rs/tuika) |
-| `CodeBlock` | ✅ stable | \
-[gallery](https://github.com/everruns/tuika/blob/main/docs/components.md) |
-| **Image**   |  🚧 beta  | \
-[features](https://github.com/everruns/tuika/blob/main/docs/features.md) |
-";
-
-/// A GFM table rendered by the one-shot `Markdown` view.
-///
-/// Column widths come from the content and are fitted to the area, so the same
-/// source reflows on resize rather than being pre-formatted by the host.
-fn scene_markdown_table(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    element(Markdown::new(MARKDOWN_TABLE_DOC))
-}
-
-const MARKDOWN_HTML_DOC: &str = "\
-### Inline HTML
-
-Markdown in the wild carries HTML, so the presentational inline
-tags render instead of being dropped — each resolving the same
-`StyleSheet` role as the markdown it mirrors.
-
-- <b>strong</b> and <em>emphasis</em>, <s>struck</s>, <u>underlined</u>
-- <mark>highlighted</mark>, <kbd>Ctrl</kbd>+<kbd>C</kbd>, <a href=\"https://docs.rs/tuika\">a link</a>
-- H<sub>2</sub>O and 3&times;10<sup>8</sup> m/s
-
-A line<br>broken by a tag.
-";
-
-/// Inline HTML inside markdown: every tag resolves a `StyleSheet` role, so
-/// `<b>` cannot look different from `**bold**`.
-fn scene_markdown_html(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    element(Markdown::new(MARKDOWN_HTML_DOC))
-}
-
-/// A single themed, syntax-highlighted fenced block via `CodeBlock`.
-fn scene_code_block(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    let source = "pub fn fib(n: u64) -> u64 {\n    match n {\n        0 | 1 => n,\n        _ => fib(n - 1) + fib(n - 2),\n    }\n}";
-    element(
-        CodeBlock::new("rust", source)
-            .highlighter(&HL)
-            .line_numbers(true),
-    )
-}
-
-fn scene_rule(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(Rule::new().style(theme.muted_style())) }
-            fixed(1) { node(Rule::new().title(Line::from(Span::styled(" Section ", theme.accent_style()))).style(theme.muted_style())) }
-            fixed(1) { node(Rule::new().glyph('┈').style(theme.muted_style())) }
-            fixed(1) { node(Rule::new().title(Line::from(Span::styled(" dotted ", theme.accent_style()))).glyph('·').style(theme.muted_style())) }
-        }
-    }
-}
-
-fn scene_boxed(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let inner = Text::new(vec![Line::from(Span::styled(
-        "border + padding + title",
-        theme.text_style(),
-    ))]);
-    let plain = Text::new(vec![Line::from(Span::styled(
-        "rounded border",
-        theme.text_style(),
-    ))]);
-    view! {
-        col(gap = 1) {
-            fixed(3) {
-                boxed(title = Line::from(Span::styled(" thick ", theme.accent_style())), border = BorderStyle::Thick, padding = Padding::symmetric(1, 0)) {
-                    node(inner)
-                }
-            }
-            fixed(3) {
-                boxed(title = Line::from(Span::styled(" rounded ", theme.accent_style())), border = BorderStyle::Rounded) {
-                    node(plain)
-                }
-            }
-        }
-    }
-}
-
-fn scene_flex(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let cell = |label: &str, color: Color| -> Element {
-        let text = Text::new(vec![Line::from(Span::styled(
-            label.to_string(),
-            Style::default()
-                .fg(theme.background)
-                .bg(color)
-                .add_modifier(Modifier::BOLD),
-        ))]);
-        element(
-            tuika::components::Boxed::new(element(text))
-                .border(BorderStyle::Plain)
-                .background(Style::default().bg(color)),
-        )
-    };
-    view! {
-        col(gap = 1) {
-            fixed(3) {
-                row(gap = 1) {
-                    grow(1) { node(cell("grow 1", theme.accent)) }
-                    grow(2) { node(cell("grow 2", theme.accent_alt)) }
-                    fixed(12) { node(cell("fixed 12", theme.muted)) }
-                }
-            }
-            fixed(1) { node(Text::new(vec![Line::from(Span::styled("row · gap 1 · grow shares leftover width", theme.muted_style()))])) }
-        }
-    }
-}
-
-fn scene_app_shell(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let query = "view";
-    let header_text = theme.text_style();
-    let header_accent = theme.accent_style();
-    let header = view_fn(
-        |available, _ctx| Size::new(available.width, available.height.min(2)),
-        move |area, surface, _ctx| {
-            surface.set_string(
-                area.x.saturating_add(2),
-                area.bottom().saturating_sub(1),
-                "Search: ",
-                header_text,
-            );
-            surface.set_string(
-                area.x.saturating_add(10),
-                area.bottom().saturating_sub(1),
-                query,
-                header_text,
-            );
-            surface.set_string(
-                area.right().saturating_sub(10),
-                area.bottom().saturating_sub(1),
-                " 3 matches ",
-                header_accent,
-            );
-        },
-    );
-    let content = Boxed::new(element(Text::new(vec![
-        Line::from(Span::styled("app_shell.rs", theme.text_style())),
-        Line::from(Span::styled("flex.rs", theme.text_style())),
-        Line::from(Span::styled("responsive.rs", theme.text_style())),
-    ])))
-    .title(" Files ");
-    let status_ready = theme.selection_style();
-    let status_muted = theme.muted_style();
-    let status = view_fn(
-        |available, _ctx| Size::new(available.width, available.height.min(1)),
-        move |area, surface, _ctx| {
-            surface.set_string(area.x, area.y, " READY ", status_ready);
-            surface.set_string(
-                area.right().saturating_sub(15),
-                area.y,
-                "borrowed state ",
-                status_muted,
-            );
-        },
-    );
-
-    element(
-        AppShell::new(content)
-            .header(header)
-            .top_rule()
-            .status(status)
-            .bottom_rule()
-            .footer(KeyHints::new([
-                ("↑/↓", "move"),
-                ("enter", "open"),
-                ("q", "quit"),
-            ])),
-    )
-}
-
-fn scene_selection_screen(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    let rows = vec![
-        Line::from("Run the requested command"),
-        Line::from("Delegate to a specialist agent"),
-        Line::from("Allow access for this command"),
-        Line::from("Allow access for this session"),
-        Line::from("Resume the interrupted task"),
-        Line::from("Start a fresh task"),
-    ];
-    let mut state = SelectState::new();
-    state.select(Some(3));
-
-    element(
-        SelectionScreen::new("Choose how to continue", rows, &state)
-            .leading_rule()
-            .trailing_rule()
-            .footer(KeyHints::new([
-                ("↑/↓", "move"),
-                ("enter", "choose"),
-                ("esc", "cancel"),
-            ])),
-    )
-}
-
-fn scene_flow(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let labels = [
-        "build",
-        "test",
-        "docs",
-        "release-ready",
-        "terminal UI",
-        "Rust",
-    ];
-    let mut flow = tuika::components::Flow::new().gap(1);
-    for label in labels {
-        flow = flow.item(element(
-            tuika::components::Boxed::new(element(Text::raw(label)))
-                .border_color(theme.accent)
-                .padding(Padding::symmetric(1, 0)),
-        ));
-    }
-    element(flow)
-}
-
-fn scene_grid(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let mut grid = tuika::components::Grid::new(3).gap(1);
-    for (label, value) in [
-        ("jobs", "12"),
-        ("passed", "12"),
-        ("failed", "0"),
-        ("time", "4.2s"),
-        ("target", "all"),
-        ("status", "ready"),
-    ] {
-        grid = grid.cell(element(
-            tuika::components::Boxed::new(element(Text::new(vec![
-                Line::from(Span::styled(label, theme.muted_style())),
-                Line::from(Span::styled(value, theme.accent_style())),
-            ])))
-            .padding(Padding::symmetric(1, 0)),
-        ));
-    }
-    element(grid)
-}
-
-fn scene_scroll(frame: u64, theme: &Theme) -> Element {
-    let lines: Vec<Line<'static>> = (1..=24)
-        .map(|i| {
-            Line::from(Span::styled(
-                format!("  line {i:>2} — content that overflows the viewport"),
-                theme.text_style(),
-            ))
-        })
-        .collect();
-    let viewport_h = 8usize;
-    let content_h = lines.len();
-    let mut state = ScrollState::new();
-    let reach = tuika::anim::ping_pong(frame, 200);
-    let steps = (reach * (content_h.saturating_sub(viewport_h) as f32 / 3.0 + 1.0)) as u32;
-    let down = Event::Mouse(Mouse::at(MouseKind::ScrollDown, 0, 0));
-    for _ in 0..steps {
-        let _ = state.handle(&down, content_h, viewport_h);
-    }
-    view! {
-        col {
-            fixed(8) { node(Scroll::new(lines, &state)) }
-        }
-    }
-}
-
-/// A transcript of *components* — bordered panels beside plain rows — scrolling
-/// by row, so a panel taller than the remaining space clips at the edge instead
-/// of snapping. This is what `Scroll` cannot do: its content is pre-wrapped
-/// lines, so anything laid out has to be flattened by hand first.
-fn scene_item_scroll(frame: u64, theme: &Theme) -> Element {
-    let items: Vec<Element> = (1..=6)
-        .map(|i| {
-            if i % 2 == 0 {
-                element(
-                    Boxed::new(element(Text::new(vec![
-                        Line::from(Span::styled(format!("panel {i}"), theme.text_style())),
-                        Line::from(Span::styled("laid out, not drawn", theme.muted_style())),
-                    ])))
-                    .title(Line::from(Span::styled(" note ", theme.accent_style()))),
-                )
-            } else {
-                element(Text::new(vec![Line::from(Span::styled(
-                    format!("  row {i} — an ordinary line of text"),
-                    theme.text_style(),
-                ))]))
-            }
-        })
-        .collect();
-    let viewport_h = 8usize;
-    let content_h = ItemScroll::measure_height(&items, 60, 1, true, &RenderCtx::new(theme));
-    let mut state = ScrollState::new();
-    let reach = tuika::anim::ping_pong(frame, 200);
-    let steps = (reach * (content_h.saturating_sub(viewport_h) as f32 / 3.0 + 1.0)) as u32;
-    let down = Event::Mouse(Mouse::at(MouseKind::ScrollDown, 0, 0));
-    for _ in 0..steps {
-        let _ = state.handle(&down, content_h, viewport_h);
-    }
-    view! {
-        col {
-            fixed(8) { node(ItemScroll::new(items, &state).gap(1)) }
-        }
-    }
-}
-
-fn scene_scrollbar(_frame: u64, theme: &Theme) -> Element {
-    let window = VirtualWindow::new(100, 24, 38);
-    let labels = Text::new(vec![
-        Line::from(Span::styled("item 38", theme.text_style())),
-        Line::from(Span::styled("   ⋮", theme.muted_style())),
-        Line::from(Span::styled("item 61", theme.text_style())),
-        Line::from(Span::styled("24 visible of 100", theme.muted_style())),
-    ]);
-    view! {
-        col(gap = 1) {
-            fixed(4) {
-                row(gap = 2) {
-                    fixed(1) { node(Scrollbar::vertical(window)) }
-                    grow(1) { node(labels) }
-                }
-            }
-            fixed(1) { node(Scrollbar::horizontal(window)) }
-        }
-    }
-}
-
-fn scene_primitives(frame: u64, theme: &Theme) -> Element {
-    let mut scroll = ScrollState::default();
-    let pan = (tuika::anim::ping_pong(frame, 160) * 14.0).round() as usize;
-    scroll.set_x_offset(pan);
-    let mut form_state = FormState::default();
-    form_state.focus(((frame / 28) % 2) as usize);
-
-    let canvas = DrawView::new(
-        |area: Rect, surface: &mut tuika::Surface<'_>, ctx: &tuika::RenderCtx<'_>| {
-            for y in area.y..area.bottom() {
-                surface.set_string(
-                    area.x,
-                    y,
-                    "界  custom grid  0123456789  →",
-                    ctx.theme.info_style(),
-                );
-            }
-        },
-    )
-    .intrinsic_size(Size::new(31, 2));
-    let preview = Viewport::new(element(canvas), Size::new(31, 2), &scroll)
-        .vertical_scrollbar(false)
-        .horizontal_scrollbar(true);
-    let form = Form::new(
-        vec![
-            FormField::new("Name", element(Text::raw("Ada"))).help("host-owned value"),
-            FormField::new("Preview", element(preview)).error("semantic validation row"),
-        ],
-        &form_state,
-    )
-    .stack_below(36);
-    let dialog = Dialog::new("Reusable primitives", element(form))
-        .size(56, 10)
-        .key_hints([("enter", "submit"), ("esc", "cancel")])
-        .dim_backdrop(true)
-        .focus_owner("primitives");
-    element(
-        Scene::new(element(Text::new(vec![Line::from(Span::styled(
-            "base screen · overlay placement resolves at render time",
-            theme.muted_style(),
-        ))])))
-        .dialog(dialog),
-    )
-}
-
-fn scene_dialog_presets(frame: u64, theme: &Theme) -> Element {
-    let phase = (frame / 48) % 4;
-    let dialog: Dialog = match phase {
-        0 => ConfirmDialog::new(
-            "Apply changes?",
-            "This will update three files in the workspace.",
-            &ConfirmDialogState::new(),
-        )
-        .confirm_label("Apply")
-        .into(),
-        1 => {
-            let mut state = ChoiceDialogState::new();
-            let down = Event::Key(Key::new(KeyCode::Down));
-            let _ = state.handle(&down, 3);
-            ChoiceDialog::new(
-                "Choose model",
-                "Select the model for the next turn.",
-                vec!["Fast".into(), "Balanced".into(), "Deep".into()],
-                &state,
-            )
-            .into()
-        }
-        2 => {
-            let mut state = MultiChoiceDialogState::new();
-            let space = Event::Key(Key::new(KeyCode::Char(' ')));
-            let _ = state.handle(&space, 3);
-            MultiChoiceDialog::new(
-                "Include context",
-                "Choose the sources sent with the request.",
-                vec![
-                    "Current file".into(),
-                    "Diagnostics".into(),
-                    "Git diff".into(),
-                ],
-                &state,
-            )
-            .into()
-        }
-        _ => InputDialog::new(
-            "Rename task",
-            "Give this activity a concise name.",
-            &InputDialogState::from_text("Review parser changes"),
-        )
-        .placeholder("Task name")
-        .into(),
-    };
-    element(
-        Scene::new(element(Text::new(vec![Line::from(Span::styled(
-            "Agent workspace · host content remains independent",
-            theme.muted_style(),
-        ))])))
-        .dialog(dialog),
-    )
-}
-
-fn scene_select(frame: u64, theme: &Theme) -> Element {
-    let items: Vec<Line<'static>> = ["Open file…", "Save", "Save As…", "Toggle theme", "Quit"]
-        .iter()
-        .map(|s| Line::from(Span::styled((*s).to_string(), theme.text_style())))
-        .collect();
-    let mut state = SelectState::new();
-    let target = (frame / 10) % items.len() as u64;
-    let down = Event::Key(Key::new(KeyCode::Down));
-    for _ in 0..target {
-        let _ = state.handle(&down, items.len());
-    }
-    view! {
-        col {
-            grow(1) { node(SelectList::new(items, &state)) }
-        }
-    }
-}
-
-struct KeyedDemoRow {
-    id: u64,
-    name: &'static str,
-    state: &'static str,
-    age: u16,
-}
-
-static KEYED_ROWS: [KeyedDemoRow; 6] = [
-    KeyedDemoRow {
-        id: 101,
-        name: "index workspace",
-        state: "running",
-        age: 4,
-    },
-    KeyedDemoRow {
-        id: 102,
-        name: "review patch",
-        state: "queued",
-        age: 12,
-    },
-    KeyedDemoRow {
-        id: 103,
-        name: "run checks",
-        state: "running",
-        age: 19,
-    },
-    KeyedDemoRow {
-        id: 104,
-        name: "refresh graph",
-        state: "queued",
-        age: 27,
-    },
-    KeyedDemoRow {
-        id: 105,
-        name: "stream logs",
-        state: "running",
-        age: 34,
-    },
-    KeyedDemoRow {
-        id: 106,
-        name: "publish report",
-        state: "queued",
-        age: 41,
-    },
-];
-
-static KEYED_REORDERED: [usize; 6] = [4, 0, 3, 2, 5, 1];
-static KEYED_ORIGINAL: [usize; 6] = [0, 1, 2, 3, 4, 5];
-static KEYED_FILTERED: [usize; 2] = [0, 4];
-static KEYED_SELECTION: KeyedSelectState<u64> = KeyedSelectState::with_selected(103);
-
-struct KeyedDemoRows {
-    visible: &'static [usize],
-}
-
-impl KeyedRowSource<u64> for KeyedDemoRows {
-    type Row = KeyedDemoRow;
-
-    fn len(&self) -> usize {
-        self.visible.len()
-    }
-
-    fn row(&self, index: usize) -> Option<&Self::Row> {
-        self.visible
-            .get(index)
-            .and_then(|&index| KEYED_ROWS.get(index))
-    }
-
-    fn key_eq(&self, _index: usize, row: &Self::Row, key: &u64) -> bool {
-        row.id == *key
-    }
-}
-
-fn keyed_demo_name(row: &KeyedDemoRow) -> Line<'_> {
-    Line::from(row.name)
-}
-
-fn keyed_demo_state(row: &KeyedDemoRow) -> Line<'_> {
-    Line::from(row.state)
-}
-
-fn keyed_demo_age(row: &KeyedDemoRow) -> Line<'_> {
-    Line::from(format!("{}s", row.age))
-}
-
-fn scene_keyed_table(frame: u64, _theme: &Theme) -> Element {
-    let phase = (frame / 18) % 3;
-    let visible: &'static [usize] = match phase {
-        0 => &KEYED_REORDERED,
-        1 => &KEYED_FILTERED,
-        _ => &KEYED_ORIGINAL,
-    };
-    let label = match phase {
-        0 => "reordered · key 103 stays selected",
-        1 => "filtered out · key 103 remains in host state",
-        _ => "stream refreshed · key 103 restored",
-    };
-    view! {
-        col {
-            fixed(1) { node(Text::raw(label)) }
-            grow(1) {
-                node(KeyedTable::source(
-                    vec![
-                        KeyedColumn::fixed("id", 5, |row: &KeyedDemoRow| Line::from(row.id.to_string())).right(),
-                        KeyedColumn::flex("task", 2, keyed_demo_name),
-                        KeyedColumn::fixed("state", 8, keyed_demo_state).hide_below(34),
-                        KeyedColumn::fixed("age", 5, keyed_demo_age).right().optional(),
-                    ],
-                    KeyedDemoRows { visible },
-                    &KEYED_SELECTION,
-                ))
-            }
-        }
-    }
-}
-
-fn scene_completion_palette(frame: u64, theme: &Theme) -> Element {
-    let _ = theme;
-    let items = vec![
-        CompletionItem::new("status")
-            .detail("Show session status")
-            .replacement("/status"),
-        CompletionItem::new("model")
-            .detail("Choose a model")
-            .replacement("/model")
-            .keyword("engine"),
-        CompletionItem::new("approvals")
-            .detail("Configure command approvals")
-            .replacement("/approvals"),
-        CompletionItem::new("review")
-            .detail("Review current changes")
-            .replacement("/review"),
-        CompletionItem::new("init")
-            .detail("Create project instructions")
-            .replacement("/init"),
-    ];
-    let queries = ["", "m", "mo", "mod"];
-    let query = queries[((frame / 24) as usize) % queries.len()];
-    let mut state = CompletionState::new();
-    state.sync(query, &items);
-    element(
-        CompletionPalette::new(&items, &state)
-            .title("Commands")
-            .viewport(5)
-            .show_query(true),
-    )
-}
-
-fn scene_tabs(frame: u64, theme: &Theme) -> Element {
-    let labels: Vec<Line<'static>> = ["Chat", "Diff", "Logs", "Files"]
-        .iter()
-        .map(|s| Line::from(Span::styled((*s).to_string(), theme.text_style())))
-        .collect();
-    let mut state = TabsState::default();
-    let target = (frame / 16) % labels.len() as u64;
-    let right = Event::Key(Key::new(KeyCode::Right));
-    for _ in 0..target {
-        let _ = state.handle(&right, labels.len());
-    }
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(Tabs::new(labels, &state)) }
-            fixed(1) { node(Text::new(vec![Line::from(Span::styled("←/→ or Tab to switch", theme.muted_style()))])) }
-        }
-    }
-}
-
-fn demo_keymap() -> Keymap<&'static str> {
-    Keymap::new()
-        .layer(
-            Layer::new("global")
-                .bind_labeled("?", "open help", "help")
-                .bind_labeled("q", "quit", "quit"),
-        )
-        .layer(
-            Layer::new("search")
-                .priority(10)
-                .bind_labeled("enter", "open result", "open")
-                .bind_labeled("ctrl+n", "next result", "next")
-                .bind_labeled("ctrl+p", "previous result", "previous"),
-        )
-}
-
-fn scene_key_hints(_frame: u64, theme: &Theme) -> Element {
-    view! {
-        col(gap = 1) {
-            node(Text::new(vec![Line::from(Span::styled("Full width", theme.accent_style()))]))
-            node(KeyHints::from_keymap(&demo_keymap()))
-            row(gap = 2) {
-                fixed(12) { node(Text::new(vec![Line::from(Span::styled("28 columns", theme.muted_style()))])) }
-                fixed(28) { node(KeyHints::from_keymap(&demo_keymap())) }
-            }
-        }
-    }
-}
-
-fn scene_keymap_help(_frame: u64, theme: &Theme) -> Element {
-    view! {
-        col(gap = 1) {
-            node(Text::new(vec![Line::from(Span::styled("Active bindings", theme.accent_style()))]))
-            node(KeymapHelp::from_keymap(&demo_keymap()))
-        }
-    }
-}
-
-fn scene_status_bar(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let bar = StatusBar::new()
-        .left(vec![
-            Span::styled(" NORMAL ", theme.selection_style()),
-            Span::styled("  main.rs", theme.text_style()),
-        ])
-        .right(vec![
-            Span::styled("utf-8  ", theme.muted_style()),
-            Span::styled("Ln 42, Col 7 ", theme.text_style()),
-        ])
-        .background(Style::default().bg(theme.surface));
-    view! {
-        col {
-            fixed(1) { node(bar) }
-        }
-    }
-}
-
-/// Typing into a composer: the placeholder while it is empty, and a `@` token
-/// colored as it is typed. The trigger table is the host's — tuika finds the
-/// token and paints the range it is handed.
-fn scene_textinput(frame: u64, theme: &Theme) -> Element {
-    let full = "fix @src/parser.rs: trailing commas";
-    let n = ((frame / 3) as usize % (full.chars().count() + 12)).min(full.chars().count());
-    let typed: String = full.chars().take(n).collect();
-    let state = TextInputState::from_text(&typed);
-    let mention = Style::default()
-        .fg(theme.code.link)
-        .add_modifier(Modifier::BOLD);
-    let highlights: Vec<TextSpan> = state
-        .tokens(&[Trigger::new('@')])
-        .iter()
-        .map(|t| t.span(mention))
-        .collect();
-    view! {
-        col {
-            fixed(3) {
-                boxed(title = Line::from(Span::styled(" commit message ", theme.accent_style()))) {
-                    node(
-                        TextInput::new(&state)
-                            .placeholder("describe the change", Style::default().fg(theme.dim))
-                            .highlights(highlights)
-                    )
-                }
-            }
-        }
-    }
-}
-
-/// Bare `http(s)` URLs the host wraps in OSC 8, and a markdown link whose label
-/// carries the target. Rendered with the theme's link color + underline — the
-/// look a supporting terminal makes clickable; others show the text unchanged.
-/// The normal paint path draws styled cells; real OSC 8 emission is the job of
-/// `HyperlinkBackend` / `write_line`, so this scene shows the *appearance*.
-fn scene_hyperlink(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let link = Style::default()
-        .fg(theme.code.link)
-        .add_modifier(Modifier::UNDERLINED);
-    let body = Text::new(vec![
-        Line::from(Span::styled(
-            "A bare URL is wrapped in place — clickable, text unchanged:",
-            theme.muted_style(),
-        )),
-        Line::from(vec![
-            Span::styled("  see ", theme.text_style()),
-            Span::styled("https://docs.rs/tuika", link),
-            Span::styled(" for the API.", theme.text_style()),
-        ]),
-        Line::from(""),
-        Line::from(Span::styled(
-            "A markdown link shows its label, hiding the target:",
-            theme.muted_style(),
-        )),
-        Line::from(vec![
-            Span::styled("  the ", theme.text_style()),
-            Span::styled("tuika component gallery", link),
-            Span::styled(" demos every widget.", theme.text_style()),
-        ]),
-        Line::from(""),
-        Line::from(Span::styled(
-            "Only http(s) links are emitted; anything else stays plain text.",
-            theme.muted_style(),
-        )),
-    ]);
-    element(body)
-}
-
-/// A left-drag selection growing over a phrase (real, copyable text), plus a
-/// row of clickable regions a `HitMap` would resolve to actions.
-fn scene_mouse(frame: u64, theme: &Theme) -> Element {
-    let phrase = "the quick brown fox jumps over the lazy dog";
-    let count = phrase.chars().count();
-    let reach = tuika::anim::ping_pong(frame, 200);
-    let selected = (reach * count as f32).round() as usize;
-    let sel: String = phrase.chars().take(selected).collect();
-    let rest: String = phrase.chars().skip(selected).collect();
-
-    let button = |label: &str, active: bool| -> Span<'static> {
-        if active {
-            Span::styled(
-                format!(" {label} "),
-                Style::default()
-                    .fg(theme.background)
-                    .bg(theme.accent)
-                    .add_modifier(Modifier::BOLD),
-            )
-        } else {
-            Span::styled(format!(" {label} "), theme.muted_style())
-        }
-    };
-    let hot = (frame / 24) % 3;
-
-    let body = Text::new(vec![
-        Line::from(Span::styled(
-            "Left-drag selects real text — copy it over SSH via OSC 52:",
-            theme.muted_style(),
-        )),
-        Line::from(vec![
-            Span::styled("  ", theme.text_style()),
-            Span::styled(sel, theme.selection_style()),
-            Span::styled(rest, theme.text_style()),
-        ]),
-        Line::from(""),
-        Line::from(Span::styled(
-            "A HitMap maps screen regions to values — clicks become actions:",
-            theme.muted_style(),
-        )),
-        Line::from(vec![
-            Span::styled("  ", theme.text_style()),
-            button("Run", hot == 0),
-            Span::styled("  ", theme.text_style()),
-            button("Diff", hot == 1),
-            Span::styled("  ", theme.text_style()),
-            button("Cancel", hot == 2),
-        ]),
-    ]);
-    element(body)
-}
-
-fn scene_overlay(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    use tuika::overlay::Extent;
-    use tuika::probe::RectProbe;
-
-    let target = RectProbe::new();
-    let base = |s: &str| {
-        Text::new(vec![Line::from(Span::styled(
-            s.to_string(),
-            theme.muted_style(),
-        ))])
-    };
-    let trigger = target.wrap(Text::new(vec![Line::from(Span::styled(
-        "[ Open actions ▾ ]",
-        theme.accent_style(),
-    ))]));
-    let root: Element = view! {
-        col(gap = 0) {
-            fixed(1) { node(base("base layer stays independently laid out")) }
-            fixed(1) { node(base("the popover follows its trigger after layout")) }
-            grow(1) { spacer() }
-            fixed(1) {
-                row {
-                    grow(1) { spacer() }
-                    fixed(20) { node(trigger) }
-                    fixed(2) { spacer() }
-                }
-            }
-            fixed(1) { node(base("preferred below · flipped above at the edge")) }
-        }
-    };
-    let popover = view! {
-        boxed(
-            title = Line::from(Span::styled(" actions ", theme.accent_style())),
-            border = BorderStyle::Rounded,
-            padding = Padding::all(1)
-        ) {
-            col(gap = 1) {
-                node(Text::new(vec![Line::from(Span::styled(
-                    "Run command",
-                    theme.text_style(),
-                ))]))
-                node(Text::new(vec![Line::from(Span::styled(
-                    "Inspect logs",
-                    theme.muted_style(),
-                ))]))
-            }
-        }
-    };
-    let spec = OverlaySpec {
-        width: Extent::Cells(28),
-        height: Extent::Cells(7),
-        ..OverlaySpec::centered(0, 0).margin(1)
-    };
-    element(
-        Scene::new(root).overlay(SceneOverlay::new(popover, spec).target(
-            &target,
-            TargetPlacement::below().align(TargetAlign::End).gap(1),
-        )),
-    )
-}
-
-/// A muted caption line, reused by several scenes below.
-fn caption(text: &str, theme: &Theme) -> Element {
-    element(Text::new(vec![Line::from(Span::styled(
-        text.to_string(),
-        theme.muted_style(),
-    ))]))
-}
-
-fn scene_diff(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    let old = "fn add(a: i32, b: i32) -> i32 {\n    a - b\n}\n\nlet total = add(2, 3);";
-    let new = "fn add(a: i32, b: i32) -> i32 {\n    a + b\n}\n\nlet total = add(2, 3);\nprintln!(\"{total}\");";
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(caption("unified · red = removed, green = added", theme)) }
-            grow(1) { node(Diff::new(old, new).line_numbers(true)) }
-        }
-    }
-}
-
-fn scene_ascii_font(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    view! {
-        col(gap = 1) {
-            fixed(5) { node(AsciiFont::new("TUIKA")) }
-            fixed(1) { node(caption("embedded 5-row block font · A-Z 0-9 punctuation", theme)) }
-        }
-    }
-}
-
-fn scene_qr(frame: u64, theme: &Theme) -> Element {
-    let _ = frame;
-    // Byte-mode v1-4 encoder; a short URL fits comfortably in version 1-2.
-    const PAYLOAD: &str = "https://everruns.com";
-    let qr = QrCode::encode(PAYLOAD, QrEcc::Medium)
-        .map(element)
-        .unwrap_or_else(|| caption("(payload too large)", theme));
-    view! {
-        col(gap = 1) {
-            grow(1) { node(qr) }
-            fixed(1) { node(caption(&format!("QrCode::encode(\"{PAYLOAD}\", QrEcc::Medium)"), theme)) }
-        }
-    }
-}
-
-fn scene_tab_select(frame: u64, theme: &Theme) -> Element {
-    let labels: Vec<Line<'static>> = ["Low", "Medium", "High", "Ultra"]
-        .iter()
-        .map(|s| Line::from(Span::styled((*s).to_string(), theme.text_style())))
-        .collect();
-    let mut state = TabSelectState::default();
-    let target = (frame / 18) % labels.len() as u64;
-    let right = Event::Key(Key::new(KeyCode::Right));
-    for _ in 0..target {
-        let _ = state.handle(&right, labels.len());
-    }
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(TabSelect::new(labels, &state)) }
-            fixed(1) { node(caption("← → move the selection · enter/space activates", theme)) }
-        }
-    }
-}
-
-fn scene_slider(frame: u64, theme: &Theme) -> Element {
-    let swept = tuika::anim::ping_pong(frame, 160);
-    let mut volume = SliderState::new(0.0, 100.0, 0.0);
-    volume.set_ratio(swept);
-    let contrast = SliderState::new(0.0, 10.0, 6.0);
-    let mut balance = SliderState::new(0.0, 1.0, 0.0);
-    balance.set_ratio(1.0 - swept);
-
-    let row = |label: &str, s: &SliderState, theme: &Theme| -> Element {
-        let lbl = Text::new(vec![Line::from(Span::styled(
-            label.to_string(),
-            theme.muted_style(),
-        ))]);
-        let slider = Slider::new(s).label(s);
-        view! {
-            row(gap = 1) {
-                fixed(9) { node(lbl) }
-                grow(1) { node(slider) }
-            }
-        }
-    };
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(row("volume", &volume, theme)) }
-            fixed(1) { node(row("contrast", &contrast, theme)) }
-            fixed(1) { node(row("balance", &balance, theme)) }
-        }
-    }
-}
-
-fn scene_timeline(frame: u64, theme: &Theme) -> Element {
-    // Three tracks over a shared 120-frame loop: a linear ramp, an eased ramp,
-    // and a 0→1→0 pulse — each a pure function of the frame counter.
-    let linear = Timeline::new()
-        .keyframe(0, 0.0)
-        .keyframe(120, 1.0)
-        .repeat(Repeat::Loop);
-    let eased = Timeline::new()
-        .keyframe(0, 0.0)
-        .ease(120, 1.0, tuika::anim::ease_in_out)
-        .repeat(Repeat::Loop);
-    let pulse = Timeline::new()
-        .keyframe(0, 0.0)
-        .keyframe(60, 1.0)
-        .keyframe(120, 0.0)
-        .repeat(Repeat::Loop);
-
-    let track = |label: &str, value: f32, theme: &Theme| -> Element {
-        let lbl = Text::new(vec![Line::from(Span::styled(
-            label.to_string(),
-            theme.muted_style(),
-        ))]);
-        view! {
-            row(gap = 1) {
-                fixed(11) { node(lbl) }
-                grow(1) { node(ProgressBar::determinate(value).percent(true)) }
-            }
-        }
-    };
-    view! {
-        col(gap = 1) {
-            fixed(1) { node(track("linear", linear.sample(frame), theme)) }
-            fixed(1) { node(track("ease_in_out", eased.sample(frame), theme)) }
-            fixed(1) { node(track("pulse", pulse.sample(frame), theme)) }
-            fixed(1) { node(caption("Timeline::new().keyframe(..).ease(..).repeat(Loop)", theme)) }
-        }
-    }
-}
-
-fn scene_toast(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    // A representative stack, newest on top. The host ticks these down; here we
-    // show the steady-state look of all four levels at once.
-    let mut toasts = Toasts::new(4);
-    toasts.push(ToastLevel::Info, "Build started");
-    toasts.push(ToastLevel::Success, "42 tests passed");
-    toasts.push(ToastLevel::Warning, "2 unused imports");
-    toasts.push(ToastLevel::Error, "Deploy to prod failed");
-    let toasts: &'static Toasts = Box::leak(Box::new(toasts));
-    view! {
-        col {
-            grow(1) { node(ToastList::new(toasts)) }
-        }
-    }
-}
-
-fn scene_console(frame: u64, theme: &Theme) -> Element {
-    let _ = (frame, theme);
-    let log = ConsoleLog::new(50);
-    for line in [
-        "INFO  server listening on 127.0.0.1:8080",
-        "DEBUG loaded 12 routes in 3ms",
-        "INFO  GET /  200  1.2ms",
-        "WARN  slow query: users.by_email  214ms",
-        "INFO  GET /health  200  0.3ms",
-        "ERROR upstream timeout after 5s",
-        "INFO  reconnecting to upstream…",
-    ] {
-        log.line(line);
-    }
-    let log: &'static ConsoleLog = Box::leak(Box::new(log));
-    view! {
-        col {
-            grow(1) { node(Console::new(log).title(" console ")) }
-        }
-    }
-}
-
-fn scene_framebuffer(frame: u64, theme: &Theme) -> Element {
-    let _ = theme;
-    let (w, h) = (56u32, 22u32);
-    let mut fb = FrameBuffer::new(w, h);
-    // A diagonal gradient background.
-    fb.shade(|x, y, _| {
-        let r = 30 + (x * 120 / w) as u8;
-        let g = 20;
-        let b = 60 + (y * 120 / h) as u8;
-        [r, g, b, 255]
-    });
-    // A "ball" bouncing over the canvas, drawn as an opaque square sprite.
-    let bx = (tuika::anim::ping_pong(frame, 120) * (w - 8) as f32) as u32;
-    let by = (tuika::anim::ping_pong(frame.wrapping_add(37), 84) * (h - 8) as f32) as u32;
-    fb.fill_rect(bx, by, 8, 8, [240, 210, 90, 255]);
-    // A scanline shader post-pass darkens every other row.
-    fb.shade(|_x, y, [r, g, b, a]| {
-        if y % 2 == 0 {
-            [r, g, b, a]
-        } else {
-            [r / 2, g / 2, b / 2, a]
-        }
-    });
-    let fb: &'static FrameBuffer = Box::leak(Box::new(fb));
-    view! {
-        col(gap = 1) {
-            grow(1) { node(FrameBufferView::new(fb, w as u16, (h / 2) as u16)) }
-            fixed(1) { node(caption("FrameBuffer → half-block cells · sprite + scanline shader", theme)) }
-        }
-    }
 }
